+++ tests/scalers_go/redis/redis_cluster_lists/redis_cluster_lists_test.go
@@ -1,7 +1,7 @@
 //go:build e2e
 // +build e2e
 
-package redis_sentinel_streams_test
+package redis_cluster_lists_test
 
 import (
 	"encoding/base64"
@@ -13,13 +13,14 @@ import (
 	"k8s.io/client-go/kubernetes"
 
 	. "github.com/kedacore/keda/v2/tests/helper"
+	redis "github.com/kedacore/keda/v2/tests/scalers_go/redis/helper"
 )
 
 // Load environment variables from .env file
 var _ = godotenv.Load("../../.env")
 
 const (
-	testName = "redis-sentinel-streams-test"
+	testName = "redis-cluster-lists-test"
 )
 
 var (
@@ -31,9 +32,10 @@ var (
 	triggerAuthenticationName = fmt.Sprintf("%s-ta", testName)
 	secretName                = fmt.Sprintf("%s-secret", testName)
 	redisPassword             = "admin"
+	redisList                 = "queue"
 	redisHost                 = fmt.Sprintf("%s-headless", testName)
-	minReplicaCount           = 1
-	maxReplicaCount           = 4
+	minReplicaCount           = 0
+	maxReplicaCount           = 2
 )
 
 type templateData struct {
@@ -48,6 +50,7 @@ type templateData struct {
 	MaxReplicaCount           int
 	RedisPassword             string
 	RedisPasswordBase64       string
+	RedisList                 string
 	RedisHost                 string
 	ItemsToWrite              int
 }
@@ -61,7 +64,7 @@ metadata:
   name: {{.DeploymentName}}
   namespace: {{.TestNamespace}}
 spec:
-  replicas: 1
+  replicas: 0
   selector:
     matchLabels:
       app: {{.DeploymentName}}
@@ -72,25 +75,23 @@ spec:
     spec:
       containers:
       - name: redis-worker
-        image: ghcr.io/kedacore/tests-redis-sentinel-streams
+        image: goku321/redis-cluster-list:v1.7
         imagePullPolicy: IfNotPresent
         command: ["./main"]
-        args: ["consumer"]
+        args: ["read"]
         env:
-        - name: REDIS_HOSTS
-          value: {{.RedisHost}}.{{.RedisNamespace}}
-        - name: REDIS_PORTS
-          value: "26379"
-        - name: REDIS_STREAM_NAME
-          value: my-stream
-        - name: REDIS_STREAM_CONSUMER_GROUP_NAME
-          value: consumer-group-1
+        - name: REDIS_ADDRESSES
+          value: {{.RedisHost}}.{{.RedisNamespace}}:6379
+        - name: LIST_NAME
+          value: {{.RedisList}}
         - name: REDIS_PASSWORD
           value: {{.RedisPassword}}
         - name: REDIS_SENTINEL_PASSWORD
           value: {{.RedisPassword}}
         - name: REDIS_SENTINEL_MASTER
           value: mymaster
+        - name: READ_PROCESS_TIME
+          value: "100"
 `
 
 	secretTemplate = `apiVersion: v1
@@ -113,9 +114,6 @@ spec:
   - parameter: password
     name: {{.SecretName}}
     key: password
-  - parameter: sentinelPassword
-    name: {{.SecretName}}
-    key: password
 `
 
 	scaledObjectTemplate = `apiVersion: keda.sh/v1alpha1
@@ -130,20 +128,13 @@ spec:
   cooldownPeriod:  10
   minReplicaCount: {{.MinReplicaCount}}
   maxReplicaCount: {{.MaxReplicaCount}}
-  advanced:
-    horizontalPodAutoscalerConfig:
-      behavior:
-        scaleDown:
-          stabilizationWindowSeconds: 15
   triggers:
-  - type: redis-sentinel-streams
+  - type: redis-cluster
     metadata:
-      hostsFromEnv: REDIS_HOSTS
-      portsFromEnv: REDIS_PORTS
-      stream: my-stream
-      consumerGroup: consumer-group-1
-      sentinelMaster: mymaster
-      pendingEntriesCount: "10"
+      addressesFromEnv: REDIS_ADDRESSES
+      listName: {{.RedisList}}
+      listLength: "5"
+      activationListLength: "10"
     authenticationRef:
       name: {{.TriggerAuthenticationName}}
 `
@@ -159,26 +150,18 @@ spec:
     spec:
       containers:
       - name: redis
-        image: ghcr.io/kedacore/tests-redis-sentinel-streams
+        image: goku321/redis-cluster-list:v1.7
         imagePullPolicy: IfNotPresent
         command: ["./main"]
-        args: ["producer"]
+        args: ["write"]
         env:
-        - name: REDIS_HOSTS
-          value: {{.RedisHost}}.{{.RedisNamespace}}
-        - name: REDIS_PORTS
-          value: "26379"
-        - name: REDIS_STREAM_NAME
-          value: my-stream
-        - name: REDIS_STREAM_CONSUMER_GROUP_NAME
-          value: consumer-group-1
+        - name: REDIS_ADDRESSES
+          value: {{.RedisHost}}.{{.RedisNamespace}}:6379
         - name: REDIS_PASSWORD
           value: {{.RedisPassword}}
-        - name: REDIS_SENTINEL_PASSWORD
-          value: {{.RedisPassword}}
-        - name: REDIS_SENTINEL_MASTER
-          value: mymaster
-        - name: NUM_MESSAGES
+        - name: LIST_NAME
+          value: {{.RedisList}}
+        - name: NO_LIST_ITEMS_TO_WRITE
           value: "{{.ItemsToWrite}}"
       restartPolicy: Never
   backoffLimit: 4
@@ -189,24 +172,35 @@ func TestScaler(t *testing.T) {
 	// Create kubernetes resources for PostgreSQL server
 	kc := GetKubernetesClient(t)
 
-	// Create Redis Sentinel
-	InstallRedisSentinel(t, kc, testName, redisNamespace, redisPassword)
+	// Create Redis Cluster
+	redis.InstallCluster(t, kc, testName, redisNamespace, redisPassword)
 
 	// Create kubernetes resources for testing
 	data, templates := getTemplateData()
 	CreateKubernetesResources(t, kc, testNamespace, data, templates)
 
+	testActivation(t, kc, data)
 	testScaleUp(t, kc, data)
 	testScaleDown(t, kc)
 
 	// cleanup
-	RemoveRedisSentinel(t, kc, testName, redisNamespace)
+	redis.RemoveCluster(t, kc, testName, redisNamespace)
 	DeleteKubernetesResources(t, kc, testNamespace, data, templates)
+}
+
+func testActivation(t *testing.T, kc *kubernetes.Clientset, data templateData) {
+	t.Log("--- testing activation ---")
+	templateTriggerJob := templateValues{"insertJobTemplate": insertJobTemplate}
+	data.ItemsToWrite = 5
+	KubectlApplyMultipleWithTemplate(t, data, templateTriggerJob)
+
+	AssertReplicaCountNotChangeDuringTimePeriod(t, kc, deploymentName, testNamespace, minReplicaCount, 60)
 }
 
 func testScaleUp(t *testing.T, kc *kubernetes.Clientset, data templateData) {
 	t.Log("--- testing scale up ---")
 	templateTriggerJob := templateValues{"insertJobTemplate": insertJobTemplate}
+	data.ItemsToWrite = 200
 	KubectlApplyMultipleWithTemplate(t, data, templateTriggerJob)
 
 	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, maxReplicaCount, 60, 3),
@@ -232,8 +226,9 @@ var data = templateData{
 	JobName:                   jobName,
 	RedisPassword:             redisPassword,
 	RedisPasswordBase64:       base64.StdEncoding.EncodeToString([]byte(redisPassword)),
+	RedisList:                 redisList,
 	RedisHost:                 redisHost,
-	ItemsToWrite:              100,
+	ItemsToWrite:              0,
 }
 
 func getTemplateData() (templateData, map[string]string) {
